Factor tip migration out into a mixin
This works around a bug in Sphinx where the print function under Python
2.7 prevents class attributes from showing up in documentation. I want
to be able to document some attributes on the Team class.
