+++ gratipay/models/team/mixins/__init__.py
@@ -1,0 +1,3 @@
+from .tip_migration import TipMigrationMixin as TipMigration
+
+__all__ = ['TipMigration']
